Work around possible bug in latest espriff core
When ESP8266WebServer.send(int, char*, char*[, int]) is called to
terminate headers the call is wrapped redirected to #send_P(int, char*,
char*, int). This implementation of #send_P uses #sendContent to send
headers instead of _currentClient.write, which adds chunking data
infront of the output, in the case of the headers violating the
protocol.

Additionally it always sends the content, even if it's empty, unlike the
other methods. This may or may not be intentional.

By calling #send(int, char*, String) instead we get a version of #send
which uses _currentClient.write and only sends the content if it's not
empty, matching the old functionality.