Fix: Set currentMode on start, led flicker
* Refresh is based on currentMode, but that's not set when loading the
config. Initially it's not set so the mode isn't found and it resets to
`Colour` and ignores the value of Mode. Setting it to `Colour` in the
led setup prevents this and allows normal initial configuration of mode
to work.
* More complex modes make the LEDs flicker; instead of running the
render code every update add a toggle to alternate between render or
show, giving other code a chance to run after between slow renders and
strip refresh. Most likely being caused by wifi processing/buffering.